+++ vkpr/infra/up/src/unix/formula/formula.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 runFormula() {
-  cp $CURRENT_PWD/vkpr.yaml "$(dirname "$0")"
+  [[ -f $CURRENT_PWD/vkpr.yaml ]] && cp $CURRENT_PWD/vkpr.yaml "$(dirname "$0")"
   rit vkpr infra start --default
 }
